+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * **BREAKING** Make `/panda` namespace of `franka_gazebo` optional
 * Add effort joint trajectory controller to be used by MoveIT
 * Make finger collisions primitive in `franka_gazebo`
+* add 'gravity_vector' gravity ROS parameter to FrankaHWSim
 
 ## 0.8.1 - 2021-09-08